+++ oci/mysql_mysql_backups_data_source.go
@@ -27,6 +27,10 @@ func MysqlMysqlBackupsDataSource() *schema.Resource {
 				Type:     schema.TypeString,
 				Required: true,
 			},
+			"creation_type": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
 			"db_system_id": {
 				Type:     schema.TypeString,
 				Optional: true,
@@ -77,6 +81,10 @@ func (s *MysqlMysqlBackupsDataSourceCrud) Get() error {
 	if compartmentId, ok := s.D.GetOkExists("compartment_id"); ok {
 		tmp := compartmentId.(string)
 		request.CompartmentId = &tmp
+	}
+
+	if creationType, ok := s.D.GetOkExists("creation_type"); ok {
+		request.CreationType = oci_mysql.BackupCreationTypeEnum(creationType.(string))
 	}
 
 	if dbSystemId, ok := s.D.GetOkExists("db_system_id"); ok {
@@ -132,6 +140,8 @@ func (s *MysqlMysqlBackupsDataSourceCrud) SetData() error {
 		}
 
 		mysqlBackup["backup_type"] = r.BackupType
+
+		mysqlBackup["creation_type"] = r.CreationType
 
 		if r.DataStorageSizeInGBs != nil {
 			mysqlBackup["data_storage_size_in_gb"] = *r.DataStorageSizeInGBs